+++ examples/web_worker_fib/build.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 if ! command -v wasm-pack 2>&1 >/dev/null;
 then
 	echo 'error: you must install wasm-pack'
