+++ examples/web_worker_fib/static/index.html
@@ -1,14 +1,12 @@
 <!doctype html>
 <html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<title>Yew web worker demo</title>
-
-		<script src="wasm.js"></script>
-		<script type="application/javascript">wasm_bindgen('wasm_bg.wasm');</script>
-	</head>
-    
-	<body>
-
-	</body>
+<head>
+    <meta charset="utf-8">
+        <title>Yew web worker demo</title>
+        <script src="wasm.js"></script>
+        <script type="application/javascript">wasm_bindgen('wasm_bg.wasm');</script>
+    </head>
+  
+    <body>
+    </body>
 </html>
